+++ resources/js/components/form/CreateContainerForm.vue
@@ -1,0 +1,11 @@
+<script setup lang="ts">
+
+</script>
+
+<template>
+<h1>Hej</h1>
+</template>
+
+<style scoped>
+
+</style>
